libsnark: split .hpp / .cpp
