+++ cpp/gadgetlib_sha256.cpp
